+++ src/swak/cli/__init__.py
@@ -12,5 +12,3 @@ __all__ = [
     'DESCRIPTION',
     'EPILOG'
 ]
-
-# ToDo: Add a "read config file" function or class for merging logic!