+++ src/swak/pt/transformer/__init__.py
@@ -1,9 +1,16 @@
-"""Custom positional encodings, self-attention and feed-forward networks."""
+"""Custom components for modern transformer architectures.
+
+These include, but are not limited to, sinusoidal anr rotary position encodings
+(RoPE), flexible grouped-query self-attention, and full-blown encoder-only
+transformer using these components.
+
+"""
 
 from .positions import Learnable, Sinusoidal, Rotary
 from .attention import MultiheadedSelfAttention, GroupedQuerySelfAttention
 from .layer import EncoderLayer
 from .encoder import Encoder
+from .compressor import Compressor
 
 __all__ = [
     'Learnable',
@@ -12,5 +19,6 @@ __all__ = [
     'GroupedQuerySelfAttention',
     'MultiheadedSelfAttention',
     'EncoderLayer',
-    'Encoder'
+    'Encoder',
+    'Compressor'
 ]